refactor: adapt to breaking changes of client-go v1

+++ internal/commands/ci/cancel/pipeline/pipeline.go
@@ -90,7 +90,7 @@ func runCancelation(
 			if err != nil {
 				return err
 			}
-			_, _, err = apiClient.Pipelines.CancelPipelineBuild(pid.ID, id)
+			_, _, err = apiClient.Pipelines.CancelPipelineBuild(pid.ID, int64(id))
 			if err != nil {
 				return err
 			}
