+++ internal/commands/cluster/agent/token/list/list_test.go
@@ -46,7 +46,7 @@ func TestList_FailToRetrieveTokens(t *testing.T) {
 
 	// setup mock expectations
 	tc.MockClusterAgents.EXPECT().
-		ListAgentTokens("OWNER/REPO", 1, nil, gomock.Any()).
+		ListAgentTokens("OWNER/REPO", int64(1), nil, gomock.Any()).
 		Return(nil, nil, errors.New("dummy API failure"))
 
 	// WHEN
@@ -72,7 +72,7 @@ func TestList_SingleToken(t *testing.T) {
 	// setup mock expectations
 	gomock.InOrder(
 		tc.MockClusterAgents.EXPECT().
-			ListAgentTokens("OWNER/REPO", 1, nil, gomock.Any()).
+			ListAgentTokens("OWNER/REPO", int64(1), nil, gomock.Any()).
 			Return([]*gitlab.AgentToken{
 				{
 					ID:              42,
@@ -86,7 +86,7 @@ func TestList_SingleToken(t *testing.T) {
 				},
 			}, nil, nil),
 		tc.MockUsers.EXPECT().
-			GetUser(100, gomock.Any(), gomock.Any()).
+			GetUser(int64(100), gomock.Any(), gomock.Any()).
 			Return(&gitlab.User{Username: "any-username"}, nil, nil),
 	)
 
@@ -119,7 +119,7 @@ func TestList_MultipleToken(t *testing.T) {
 	// setup mock expectations
 	gomock.InOrder(
 		tc.MockClusterAgents.EXPECT().
-			ListAgentTokens("OWNER/REPO", 1, nil, gomock.Any()).
+			ListAgentTokens("OWNER/REPO", int64(1), nil, gomock.Any()).
 			Return([]*gitlab.AgentToken{
 				{
 					ID:              42,
@@ -143,7 +143,7 @@ func TestList_MultipleToken(t *testing.T) {
 				},
 			}, nil, nil),
 		tc.MockUsers.EXPECT().
-			GetUser(100, gomock.Any(), gomock.Any()).
+			GetUser(int64(100), gomock.Any(), gomock.Any()).
 			Return(&gitlab.User{Username: "any-username"}, nil, nil),
 	)
