+++ internal/commands/cluster/agent/update_kubeconfig/agent_update_kubeconfig.go
@@ -120,7 +120,7 @@ func (o *options) run() error {
 
 	// Retrieve agent information, most importantly its name to use it as context name.
 	repoFullName := repo.FullName()
-	agent, _, err := client.ClusterAgents.GetAgent(repoFullName, int(o.agentID)) // FIXME remove cast
+	agent, _, err := client.ClusterAgents.GetAgent(repoFullName, o.agentID)
 	if err != nil {
 		return err
 	}
@@ -207,7 +207,7 @@ func updateKubeconfig(params updateKubeconfigParams) (clientcmdapi.Config, strin
 		params.glabExecutable,
 		params.glHost,
 		params.glRepoFullName,
-		int64(params.agent.ID), // FIXME remove cast
+		params.agent.ID,
 		params.tokenExpiryDuration,
 		params.cacheMode,
 		params.checkRevoked,
