+++ internal/commands/milestone/get/get_test.go
@@ -41,7 +41,7 @@ func Test_GetProjectMilestone(t *testing.T) {
 			ExpectedMsg: []string{"Title: Milestone title\nDescription: Example description\nState: closed\nDue Date: 2025-01-15\n\n"},
 			cli:         "123 --project 456",
 			setupMock: func(tc *gitlabtesting.TestClient) {
-				tc.MockMilestones.EXPECT().GetMilestone("456", 123).Return(testMilestone, nil, nil)
+				tc.MockMilestones.EXPECT().GetMilestone("456", int64(123)).Return(testMilestone, nil, nil)
 			},
 		},
 		{
@@ -49,7 +49,7 @@ func Test_GetProjectMilestone(t *testing.T) {
 			wantErr: true,
 			cli:     "111 --project 456",
 			setupMock: func(tc *gitlabtesting.TestClient) {
-				tc.MockMilestones.EXPECT().GetMilestone("456", 111).Return(nil, nil, errors.New("404 Not found"))
+				tc.MockMilestones.EXPECT().GetMilestone("456", int64(111)).Return(nil, nil, errors.New("404 Not found"))
 			},
 		},
 	}
@@ -108,7 +108,7 @@ func Test_GetGroupMilestone(t *testing.T) {
 			ExpectedMsg: []string{"Title: Milestone title\nDescription: Example description\nState: closed\nDue Date: 2025-01-15\n\n"},
 			cli:         "123 --group 456",
 			setupMock: func(tc *gitlabtesting.TestClient) {
-				tc.MockGroupMilestones.EXPECT().GetGroupMilestone(gomock.Any(), 123).Return(testMilestone, nil, nil)
+				tc.MockGroupMilestones.EXPECT().GetGroupMilestone(gomock.Any(), int64(123)).Return(testMilestone, nil, nil)
 			},
 		},
 		{
@@ -116,7 +116,7 @@ func Test_GetGroupMilestone(t *testing.T) {
 			wantErr: true,
 			cli:     "111 --group 456",
 			setupMock: func(tc *gitlabtesting.TestClient) {
-				tc.MockGroupMilestones.EXPECT().GetGroupMilestone(gomock.Any(), 111).Return(nil, nil, errors.New("404 Not found"))
+				tc.MockGroupMilestones.EXPECT().GetGroupMilestone(gomock.Any(), int64(111)).Return(nil, nil, errors.New("404 Not found"))
 			},
 		},
 	}
